requirements.txt
anyio==3.3.4
asgiref==3.4.1
click==8.0.3
fastapi==0.70.0
Flask==2.0.2
googleapis-common-protos==1.53.0
grpcio==1.41.1
grpcio-tools==1.41.1
h11==0.12.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
protobuf==3.19.0
pydantic==1.8.2
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
Werkzeug==2.0.2

pytest~=7.1.3